# expression: implement vectorized evaluation for `builtinCastDecimalAsDecimalSig`