PCP #12102

## What have you changed?

implement vectorized evaluation for `builtinCastDecimalAsDecimalSig`

## What is changed and how it works?
```bash
$$ go test -check.f TestVectorizedBuiltinCastFunc
PASS: builtin_cast_vec_test.go:130: testEvaluatorSuite.TestVectorizedBuiltinCastFunc    0.146s
OK: 1 passed
PASS
ok      github.com/pingcap/tidb/expression      0.166s

$$ go test -v -benchmem -bench=BenchmarkVectorizedBuiltinCastFunc -run=BenchmarkVectorizedBuiltinCastFunc -args "builtinCastDecimalAsDecimalSig"
goos: darwin
goarch: amd64
pkg: github.com/pingcap/tidb/expression
BenchmarkVectorizedBuiltinCastFunc/builtinCastDecimalAsDecimalSig-VecBuiltinFunc-8                 69488             16760 ns/op               0 B/op             0 allocs/op
BenchmarkVectorizedBuiltinCastFunc/builtinCastDecimalAsDecimalSig-NonVecBuiltinFunc-8              19612             61114 ns/op           40032 B/op           834 allocs/op
PASS
ok      github.com/pingcap/tidb/expression      3.200s

```

## Check List

Tests
  - Unit test